added demo user

+++ src/cmps/reservation-preview.jsx
@@ -1,16 +1,18 @@
 import { utilService } from "../services/util.service"
 import { reservationService } from "../services/reservation.service"
 
-export function ReservationPreview({ reservation, getReservations }) {
+export function ReservationPreview({ reservation, getReservations, loggedInUser }) {
 
     const onRemove = async () => {
+        if (loggedInUser.username === 'demo') {
+            alert('demo user cannot reject reservations.')
+            return
+        }
         await reservationService.removeReservation(reservation)
         getReservations()
     }
 
     return (<>
-       
-       
         <td style={{ textAlign: 'left' }}>
             <span style={{ marginLeft: '4px' }}>{reservation.stay.name}</span>
         </td>
@@ -19,10 +21,10 @@ export function ReservationPreview({ reservation, getReservations }) {
             <tr>
                 <td>{reservation.user.fullName}</td>
             </tr>
-            {reservation.adults!=0 && <tr>
+            {reservation.adults != 0 && <tr>
                 <td>Adults: {reservation.adults}</td>
             </tr>}
-            {reservation.childrens!=0 &&  <tr>
+            {reservation.childrens != 0 && <tr>
                 <td>Children: {reservation.childrens}</td>
             </tr>}
         </td>
@@ -41,7 +43,7 @@ export function ReservationPreview({ reservation, getReservations }) {
         <td>
             ${utilService.getUsPrice(reservation.totalPrice)}
         </td>
-        
+
         <td>
             <button className="clickable" onClick={onRemove}>Reject</button>
         </td>
